pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mermaidpaper"
version = "0.1.0"
description = "Detect and render Mermaid in PDFs/DOCs"
dependencies = ["fastapi", "uvicorn", "PyMuPDF", "python-docx", "pytest", "cairosvg", "Pillow"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]